chore: rename services/logger to services/communications
